backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pandas>=2.0.0
gspread>=5.12.0
google-auth>=2.22.0
google-api-python-client>=2.100.0
matplotlib>=3.7.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
pymupdf>=1.23.0
holidays>=0.44
